Currently, EqualsUsing recurses with Equals, which loses the allocator that you're using, defeating the whole point. Address that in all places I could find. I don't think that code-path is specifically used anywhere so this has no impact on benchmarks.